Struct alloy_rpc_client::builder::ClientBuilder

pub struct ClientBuilder<L> {
    pub(crate) builder: ServiceBuilder<L>,
}

A builder for the transport RpcClient.

This is a wrapper around [tower::ServiceBuilder]. It allows you to configure middleware layers that will be applied to the transport, and has some shortcuts for common layers and transports.

A builder accumulates Layers, and then is finished via the ClientBuilder::connect method, which produces an RPC client.

Fields
builder: ServiceBuilder<L>
Implementations

impl<L> ClientBuilder<L>

pub fn layer<M>(self, layer: M) -> ClientBuilder<Stack<M, L>>

Add a middleware layer to the stack.

This is a wrapper around [tower::ServiceBuilder::layer]. Layers that are added first will be called with the request first.

pub fn transport<T>(self, transport: T, is_local: bool) -> RpcClient<L::Service>
where L: Layer<T>, T: Transport, L::Service: Transport,

Create a new RpcClient with the given transport and the configured layers.

pub fn reqwest_http(self, url: Url) -> RpcClient<L::Service>
where L: Layer<Http<Client>>, L::Service: Transport,

Available on crate feature reqwest only.

Convenience function to create a new RpcClient with a [reqwest] HTTP transport.

pub fn hyper_http(self, url: Url) -> RpcClient<L::Service>
where L: Layer<Http<HyperClient>>, L::Service: Transport,

Available on non-WebAssembly and crate feature hyper only.

Convenience function to create a new RpcClient with a hyper HTTP transport.

pub async fn pubsub<C>( self, pubsub_connect: C ) -> TransportResult<RpcClient<L::Service>>
where C: PubSubConnect, L: Layer<PubSubFrontend>, L::Service: Transport,

Available on crate feature pubsub only.

Connect a pubsub transport, producing an RpcClient with the provided connection.

pub async fn ws( self, ws_connect: WsConnect ) -> TransportResult<RpcClient<L::Service>>
where L: Layer<PubSubFrontend>, L::Service: Transport,

Available on crate feature ws only.

Connect a WS transport, producing an RpcClient with the provided connection

pub async fn ipc<T>( self, ipc_connect: IpcConnect<T> ) -> TransportResult<RpcClient<L::Service>>
where IpcConnect<T>: PubSubConnect, L: Layer<PubSubFrontend>, L::Service: Transport,

Available on crate feature ipc only.

Connect an IPC transport, producing an RpcClient with the provided connection.

pub async fn connect<C>( self, connect: C ) -> TransportResult<RpcClient<L::Service>>
where C: TransportConnect, L: Layer<C::Transport>, L::Service: Transport,

Connect a transport, producing an RpcClient with the provided connection.

pub async fn connect_boxed<C>( self, connect: C ) -> TransportResult<RpcClient<L::Service>>
where C: BoxTransportConnect, L: Layer<BoxTransport>, L::Service: Transport,

Connect a transport, producing an RpcClient with a [BoxTransport] connection.