Enum alloy_rpc_client::builtin::BuiltInConnectionString

#[non_exhaustive]
pub enum BuiltInConnectionString {
    Http(Url),
    Ws(Url, Option<Authorization>),
    Ipc(String),
}

Connection string for built-in transports.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Http(Url)

Available on crate features reqwest or hyper only.

HTTP transport.

Ws(Url, Option<Authorization>)

Available on crate feature ws only.

WebSocket transport.

Ipc(String)

Available on crate feature ipc only.

IPC transport.

Implementations

impl BuiltInConnectionString

pub async fn connect_boxed(&self) -> Result<BoxTransport, TransportError>

Connect with the given connection string.

Notes

If hyper feature is enabled

WS will extract auth, however, auth is disabled for wasm.

pub fn try_as_http(s: &str) -> Result<Self, TransportError>

Available on crate features reqwest or hyper only.

Tries to parse the given string as an HTTP URL.

pub fn try_as_ws(s: &str) -> Result<Self, TransportError>

Available on crate feature ws only.

Tries to parse the given string as a WebSocket URL.

pub fn try_as_ipc(s: &str) -> Result<Self, TransportError>

Available on crate feature ipc only.

Tries to parse the given string as an IPC path, returning an error if the path does not exist.