Struct alloy_rpc_client::client::RpcClientInner

pub struct RpcClientInner<T> {
    pub(crate) transport: T,
    pub(crate) is_local: bool,
    pub(crate) id: AtomicU64,
}

A JSON-RPC client.

This struct manages a [Transport] and a request ID counter. It is used to build RpcCall and BatchRequest objects. The client delegates transport access to the calls.

Note

IDs are allocated sequentially, starting at 0. IDs are reserved via RpcClientInner::next_id. Note that allocated IDs may not be used. There is no guarantee that a prepared RpcCall will be sent, or that a sent call will receive a response.

Fields
transport: T

The underlying transport.

is_local: bool

true if the transport is local.

id: AtomicU64

The next request ID to use.

Implementations

impl RpcClientInner<PubSubFrontend>

pub async fn get_raw_subscription(&self, id: U256) -> RawSubscription

Available on crate feature pubsub only.

Get a [RawSubscription] for the given subscription ID.

pub async fn get_subscription<T: DeserializeOwned>( &self, id: U256 ) -> Subscription<T>

Available on crate feature pubsub only.

Get a [Subscription] for the given subscription ID.

impl<T> RpcClientInner<T>

pub const fn new(t: T, is_local: bool) -> Self

Create a new RpcClient with the given transport.

pub const fn default_poll_interval(&self) -> Duration

Returns the default poll interval for the client.

pub const fn transport(&self) -> &T

Returns a reference to the underlying transport.

pub fn transport_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying transport.

pub fn into_transport(self) -> T

Consumes the client and returns the underlying transport.

pub fn make_request<Params: RpcParam>( &self, method: impl Into<Cow<'static, str>>, params: Params ) -> Request<Params>

Build a JsonRpcRequest with the given method and params.

This function reserves an ID for the request, however the request is not sent.

To send a request, use RpcClientInner::request and await the returned RpcCall.

pub const fn is_local(&self) -> bool

true if the client believes the transport is local.

This can be used to optimize remote API usage, or to change program behavior on local endpoints. When the client is instantiated by parsing a URL or other external input, this value is set on a best-efforts basis and may be incorrect.

pub fn set_local(&mut self, is_local: bool)

Set the is_local flag.

fn increment_id(&self) -> u64

Reserve a request ID value. This is used to generate request IDs.

pub fn next_id(&self) -> Id

Reserve a request ID u64.

impl<T: Transport + Clone> RpcClientInner<T>

pub fn request<Params: RpcParam, Resp: RpcReturn>( &self, method: impl Into<Cow<'static, str>>, params: Params ) -> RpcCall<T, Params, Resp>

Prepares an RpcCall.

This function reserves an ID for the request, however the request is not sent. To send a request, await the returned RpcCall.

Note

Serialization is done lazily. It will not be performed until the call is awaited. This means that if a serializer error occurs, it will not be caught until the call is awaited.

pub fn boxed(self) -> RpcClientInner<BoxTransport>

Type erase the service in the transport, allowing it to be used in a generic context.

Note:

This is for abstracting over RpcClient<T> for multiple T by erasing each type. E.g. if you have RpcClient<Http> and RpcClient<Ws> you can put both into a Vec<RpcClient<BoxTransport>>.