Struct alloy_rpc_client::RpcClient

pub struct RpcClient<T>(Arc<RpcClientInner<T>>);

A JSON-RPC client.

RpcClient should never be instantiated directly. Instead, use ClientBuilder.

Tuple Fields
0: Arc<RpcClientInner<T>>
Implementations

impl RpcClient<PubSubFrontend>

pub async fn connect_pubsub<C>( connect: C ) -> TransportResult<RpcClient<PubSubFrontend>>
where C: PubSubConnect,

Available on crate feature pubsub only.

Connect to a transport via a [PubSubConnect] implementor.

pub fn channel_size(&self) -> usize

Available on crate feature pubsub only.

Get the currently configured channel size. This is the number of items to buffer in new subscription channels. Defaults to 16. See tokio::sync::broadcast for a description of relevant behavior.

impl RpcClient<Identity>

pub fn builder() -> ClientBuilder<Identity>

Create a new ClientBuilder.

impl RpcClient<Http<Client>>

pub fn new_http(url: Url) -> Self

Available on crate feature reqwest only.

Create a new RpcClient with an HTTP transport.

impl<T> RpcClient<T>

pub fn new(t: T, is_local: bool) -> Self

Creates a new RpcClient with the given transport.

pub fn from_inner(inner: RpcClientInner<T>) -> Self

Creates a new RpcClient with the given inner client.

pub const fn inner(&self) -> &Arc<RpcClientInner<T>>

Get a reference to the client.

pub fn into_inner(self) -> Arc<RpcClientInner<T>>

Convert the client into its inner type.

pub fn get_weak(&self) -> WeakClient<T>

Get a Weak reference to the client.

pub fn get_ref(&self) -> ClientRef<'_, T>

Borrow the client.

impl<T: Transport> RpcClient<T>

pub fn prepare_static_poller<Params, Resp>( &self, method: impl Into<Cow<'static, str>>, params: Params ) -> PollerBuilder<T, Params, Resp>
where T: Clone, Params: RpcParam + 'static, Resp: RpcReturn + Clone,

Build a poller that polls a method with the given parameters.

See PollerBuilder for examples and more details.

impl<T: Transport + Clone> RpcClient<T>

pub fn boxed(self) -> RpcClient<BoxTransport>

Boxes the transport.

This will create a new client if this instance is not the only reference to the inner client.

impl<T> RpcClient<Http<T>>

pub fn new_batch(&self) -> BatchRequest<'_, Http<T>>

Create a new BatchRequest builder.